+++ .storybook/preview.ts
@@ -1,7 +1,6 @@
 import { withPerformance } from 'storybook-addon-performance';
 import type { GlobalTypes } from '@storybook/csf';
 import type { DecoratorFn, Parameters } from '@storybook/react';
-import { systemPropNames } from '@vibrant-ui/core';
 import { withGlobalStyle } from './decorators/withGlobalStyle';
 import { withTheme } from './decorators/withTheme';
 
@@ -11,7 +10,6 @@ export const parameters: Parameters = {
     argTypesRegex: '^on[A-Z].*',
   },
   controls: {
-    exclude: systemPropNames,
     sort: 'requiredFirst',
     expanded: true,
   },