+++ packages/vibrant-theme/src/lib/types/Theme.ts
@@ -1,3 +1,4 @@
+import type { ThemeBorderRadius } from './BorderRadius';
 import type { Colors, ThemeColors } from './Colors';
 import type { ThemeGradient } from './Gradient';
 import type { ThemeMode } from './Mode';
@@ -9,6 +10,7 @@ export type Theme = {
   colors: ThemeColors;
   opacity: ThemeOpacity;
   gradient: ThemeGradient;
+  borderRadius: ThemeBorderRadius;
   typography: ThemeTypography;
   typographyWeight: ThemeTypographyWeight;
   mode: ThemeMode;
@@ -18,6 +20,7 @@ export type CurrentTheme = {
   breakpoints: number[];
   colors: Colors;
   opacity: Opacity;
+  borderRadius: ThemeBorderRadius;
   typography: ThemeTypography;
   typographyWeight: ThemeTypographyWeight;
 };