+++ packages/utils-eslint-config/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [0.2.1](https://github.com/pedaling/opensource/compare/utils-eslint-config-0.2.0...utils-eslint-config-0.2.1) (2022-07-27)
+
 ## [0.2.0](https://github.com/pedaling/opensource/compare/utils-eslint-config-0.1.1...utils-eslint-config-0.2.0) (2022-07-19)
 
 ### [0.1.1](https://github.com/pedaling/opensource/compare/utils-eslint-config-0.1.0...utils-eslint-config-0.1.1) (2022-07-19)