+++ packages/vibrant-motion/CHANGELOG.md
@@ -2,6 +2,28 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.3.0](https://github.com/pedaling/opensource/compare/vibrant-motion-0.2.0...vibrant-motion-0.3.0) (2022-07-27)
+
+### Dependency Updates
+
+* `vibrant-core` updated to version `0.4.0`
+* `vibrant-utils` updated to version `0.3.1`
+* `vibrant-theme` updated to version `0.3.0`
+
+### Features
+
+* motion 컴포넌트가 ResponsiveValue로 동작되도록 한다 ([#106](https://github.com/pedaling/opensource/issues/106)) ([e15b8fb](https://github.com/pedaling/opensource/commit/e15b8fb2ccca804aa7069faf85165e0ede8ac1e0))
+* **vibrant-components:** pressable prop을 제거하고 Pressable 컴포넌트를 추가한다 ([#110](https://github.com/pedaling/opensource/issues/110)) ([a4246cd](https://github.com/pedaling/opensource/commit/a4246cd525c7cd8910db1d495d40e0802f74a5fa))
+* **vibrant-motion:** add backgroundColor prop to Motion ([#83](https://github.com/pedaling/opensource/issues/83)) ([7cbbfdf](https://github.com/pedaling/opensource/commit/7cbbfdfaa1048951a76873cd93c057ac93d99e12))
+* **vibrant-motion:** add Transition Component ([#94](https://github.com/pedaling/opensource/issues/94)) ([231e616](https://github.com/pedaling/opensource/commit/231e61623c4ad3065f62f6f26b60f0927b2790f4))
+
+
+### Bug Fixes
+
+* motion을 ref로 컨트롤할 수 있도록 한다 ([#89](https://github.com/pedaling/opensource/issues/89)) ([ed2479e](https://github.com/pedaling/opensource/commit/ed2479ecd0b871c673b2ddf6285082da72ea21d4))
+* update native jest config to use jest-expo preset ([#109](https://github.com/pedaling/opensource/issues/109)) ([0ef07f1](https://github.com/pedaling/opensource/commit/0ef07f108156cffd0e4cdb535d6689dfeb8f9463))
+* **vibrant-motion:** fix native transform motion ([#84](https://github.com/pedaling/opensource/issues/84)) ([2b87295](https://github.com/pedaling/opensource/commit/2b872951a052906d9590f07f95cbc92545d91f39))
+
 ## [0.2.0](https://github.com/pedaling/opensource/compare/vibrant-motion-0.1.2...vibrant-motion-0.2.0) (2022-07-19)
 
 ### [0.1.2](https://github.com/pedaling/opensource/compare/vibrant-motion-0.1.1...vibrant-motion-0.1.2) (2022-07-19)