+++ packages/vibrant-theme/CHANGELOG.md
@@ -2,6 +2,25 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.3.0](https://github.com/pedaling/opensource/compare/vibrant-theme-0.2.0...vibrant-theme-0.3.0) (2022-07-27)
+
+### Dependency Updates
+
+* `vibrant-utils` updated to version `0.3.1`
+
+### Features
+
+* opacity Token을 추가한다 ([#105](https://github.com/pedaling/opensource/issues/105)) ([cd5c406](https://github.com/pedaling/opensource/commit/cd5c4067b08d22d4d0c68c482b95528896aa6738))
+* **vibrant-theme, vibrant-components:** gradient scale을 theme에 추가한다 ([#91](https://github.com/pedaling/opensource/issues/91)) ([5a9c70b](https://github.com/pedaling/opensource/commit/5a9c70b4013e44534b536c25c5ee7769a0ead55e))
+* **vibrant-theme:** add borderRadius theme value ([#118](https://github.com/pedaling/opensource/issues/118)) ([c6c6ccd](https://github.com/pedaling/opensource/commit/c6c6ccda83ecb813c37cb979f88bfeaa2aa0efe0))
+* **vibrant-theme:** add transparent to ColorToken ([#93](https://github.com/pedaling/opensource/issues/93)) ([94f3041](https://github.com/pedaling/opensource/commit/94f30410c1a53c36beb6dd973f7ed10d2e4825c5))
+* **vibrant-theme:** surface4, white color 추가 ([#120](https://github.com/pedaling/opensource/issues/120)) ([c4de17b](https://github.com/pedaling/opensource/commit/c4de17ba700a9c24aae2ba1f021a89fb254a7605))
+
+
+### Bug Fixes
+
+* **vibrant-motion:** fix native transform motion ([#84](https://github.com/pedaling/opensource/issues/84)) ([2b87295](https://github.com/pedaling/opensource/commit/2b872951a052906d9590f07f95cbc92545d91f39))
+
 ## [0.2.0](https://github.com/pedaling/opensource/compare/vibrant-theme-0.1.1...vibrant-theme-0.2.0) (2022-07-19)
 
 ### [0.1.1](https://github.com/pedaling/opensource/compare/vibrant-theme-0.1.0...vibrant-theme-0.1.1) (2022-07-19)